requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
Django==5.0.6
django-celery-results==2.5.1
djangorestframework==3.15.1
filelock==3.14.0
identify==2.5.36
idna==3.7
kombu==5.3.7
nodeenv==1.8.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
PyYAML==6.0.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
